+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+Version 1.2 - 16th September 2012
+=================================
+
+Add support for photo upload and replace.
+
+Version 1.1 - 16th September 2012
+=================================
+
+Restructure code to place in DPZ namespace.
+
+
 Version 1.0 - 4th August 2012
 =============================